bk: 677a08d4z6dy9prIaLZzJOyrqK2E-w

+++ include/ntp_random.h
@@ -4,7 +4,7 @@
 void ntp_crypto_srandom(void);
 int ntp_crypto_random_buf(void *buf, size_t nbytes);
 
-long ntp_random (void);
+int32 ntp_random(void);
 double ntp_uurandom(void);
 void ntp_srandom (unsigned long);
 void ntp_srandomdev (void);
